app/requirements.txt
databricks-sdk>=0.50.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
